add tests for keyboard edits
